+++ event.go
@@ -11,10 +11,3 @@ type FileEvent struct {
 
   fileinfo *os.FileInfo
 }
-
-type FileState struct {
-  Source *string `json:"source,omitempty"`
-  Offset int64 `json:"offset,omitempty"`
-  Inode uint64 `json:"inode,omitempty"`
-  Device int32 `json:"device,omitempty"`
-}
